requirements.txt
atomicwrites==1.3.0
attrs==19.2.0
certifi==2019.9.11
chardet==3.0.4
coverage==4.5.4
Django>=2.2.10
django-cors-headers==3.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
Faker==0.9.1
httpie==1.0.3
idna==2.8
importlib-metadata==0.23
migrate==0.3.8
mixer==6.1.3
more-itertools==7.2.0
packaging==19.2
pluggy==0.13.0
py==1.8.0
pycodestyle==2.5.0
Pygments==2.4.2
PyJWT==1.7.1
pyparsing==2.4.2
pytest==5.2.1
pytest-cov==2.8.1
pytest-django==3.5.1
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
urllib3==1.25.6
wcwidth==0.1.7
zipp==0.6.0
